requirements.txt
numpy==2.1.2
pandas==2.2.3
scikit_learn==1.5.2
tensorflow==2.10.0
tensorflow_intel==2.18.0
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
